Add instructions for downloading release images
We had problems rebuilding the image from master because the release was
more than a year ago.  It is not always possible to build an image from
scratch even if the docker config hasn't changed.

We should include an image for every release so it's possible to use the
saved image instead of building a new one every time.