Shutdown jobs queue and stop kue app server on sails.lower
We use "kue" to create a queue for html5 page conversions.  There is a
server listening on port 3000 in bootstrap.js when sails starts up, but
there is no code to shutdown the server when sails is shutdown.

This adds a hook on `sails.lower` to shutdown the jobs queue and the kue
app server.