Add code to start sails server in tests
Jasmine and cucumber tests expect a server to be running.  It's best to
include the code to start and stop the server in the test set up code so
it is easy to run.  For jasmine tests, we add event handlers for
jasmineStarted and jasmineDone events.  For cucumber tests, we add
`BeforeAll` and `AfterAll` handlers.